+++ app/api/orgs/[orgNameOrId]/status/route.ts
@@ -1,12 +1,13 @@
 import { auth } from 'auth'
 import { ReasonPhrases, StatusCodes } from 'http-status-codes'
 import { UserOrgStatus } from 'lib/data-model-types'
+import { getBoostOrgStatus } from 'lib/polyverse/backend/backend'
 import { NextAuthRequest } from 'next-auth/lib'
 
-import { getBoostOrgUserStatus } from './../../../../../../../lib/polyverse/backend/backend'
-import getOrg from './../../../../../../../lib/polyverse/db/get-org'
-import getUser from './../../../../../../../lib/polyverse/db/get-user'
-import authz from './../../../../../../../app/api/authz'
+import getOrg from '../../../../../lib/polyverse/db/get-org'
+import getUser from '../../../../../lib/polyverse/db/get-user'
+
+import authz from '../../../authz'
 
 export const GET = auth(async (req: NextAuthRequest) => {
   const { auth } = req
@@ -25,49 +26,47 @@ export const GET = auth(async (req: NextAuthRequest) => {
     const reqUrl = new URL(req.url)
     const reqUrlSlices = reqUrl.toString().split('/')
 
-    // The 4th to the last slice ought to be the slug for the org id
-    const requestedOrgId = reqUrlSlices[reqUrlSlices.length - 4]
+    // The 2nd to the last slice ought to be the slug for the org name
+    const requestedOrgName = reqUrlSlices[reqUrlSlices.length - 2]
 
-    const org = await getOrg(requestedOrgId)
     const user = await getUser(auth.user.email)
 
-    try {
-      authz.userListedOnOrg(org, user.id)
-      authz.orgListedOnUser(user, org.id)
-    } catch (error) {
-      return new Response(ReasonPhrases.FORBIDDEN, {
-        status: StatusCodes.FORBIDDEN,
-      })
-    }
-
-    const boostOrgUserStatus = await getBoostOrgUserStatus(org.name, user.email)
+    // try {
+    //   authz.userListedOnOrg(org, user.id)
+    //   authz.orgListedOnUser(user, org.id)
+    // } catch (error) {
+    //   return new Response(ReasonPhrases.FORBIDDEN, {
+    //     status: StatusCodes.FORBIDDEN,
+    //   })
+    // }
+    const boostOrgStatus = await getBoostOrgStatus(requestedOrgName, user.email)
 
     // Convert the response format from the Boost Node backend to something we
     // expect or consistent with our developer experience (DX).
-    const orgUserStatus: UserOrgStatus = {
+    const orgStatus: UserOrgStatus = {
       // If the `username` data member shows up on the user status that means
-      // the user has the GitHub app installed.
+      // the org has the GitHub app installed.
       gitHubAppInstalled:
-        boostOrgUserStatus.backgroundAnalysisAuthorized !== undefined &&
-        boostOrgUserStatus.backgroundAnalysisAuthorized
+        boostOrgStatus.backgroundAnalysisAuthorized !== undefined &&
+        boostOrgStatus.backgroundAnalysisAuthorized
           ? 'INSTALLED'
           : 'UNKNOWN',
 
       isPremium:
-        boostOrgUserStatus.plan && boostOrgUserStatus.plan === 'premium'
+        boostOrgStatus.plan && boostOrgStatus.plan === 'premium'
           ? 'PREMIUM'
           : 'FREE',
     }
 
-    return new Response(JSON.stringify(orgUserStatus), {
+    return new Response(JSON.stringify(orgStatus), {
       status: StatusCodes.OK,
     })
   } catch (error) {
     console.error(
-      `Failed to get user org status for '${auth.user.username}' because: ${error}`,
+      `Failed to get org status for '${auth.user.username}' because: ${error}`,
     )
 
-    return new Response('Failed to get user org status', {
+    return new Response('Failed to get org status', {
       status: StatusCodes.INTERNAL_SERVER_ERROR,
     })
   }